Go back to 32 bit intermediary values

+++ src/bm/nnue.rs
@@ -65,7 +65,7 @@ impl Nnue {
             bytes,
         ));
         bytes = &bytes[MID_0 * MID_1 * 2..];
-        let hidden_bias = include::bias_from_bytes_i16::<i16, MID_1>(bytes);
+        let hidden_bias = include::bias_from_bytes_i16::<i32, MID_1>(bytes);
         bytes = &bytes[MID_1 * 2..];
 
         let out = Arc::new(*include::dense_from_bytes_i8::<i8, MID_1, OUTPUT>(bytes));
